added findElement and a special webelement locator
findElement returns a webelement=X locator which can be used in other methods to directly execute operators on this WebElement. This opens up the possibility for chaining calls, or even relative element finding.
The retrieved webelements via findElement are cached in the WebDriverHelper using an LRU map with max size 64.
Note, webelements can become invalid after performing operations on these webelements result in failures. This of course can also be used to your advantage to check if an element was replaced using:
| reject | present | $webElementReference |
This can come in handy with AJAX calls which replace content.